refactor: Major refactor to support flexibly moving buttons across components
