+++ package.json
@@ -22,9 +22,17 @@
       "import": "./dist/index.js",
       "require": "./dist/index.cjs"
     },
-    "./dist/react-chatbotify.css": {
-      "import": "./dist/react-chatbotify.css",
-      "require": "./dist/react-chatbotify.css"
+    "./constants": {
+      "import": "./dist/constants",
+      "require": "./dist/constants"
+    },
+    "./context": {
+      "import": "./dist/context",
+      "require": "./dist/context"
+    },
+    "./types": {
+      "import": "./dist/types",
+      "require": "./dist/types"
     }
   },
   "scripts": {
