+++ src/components/Buttons/AudioButton/AudioButton.css
@@ -1,33 +1,5 @@
-/* ChatBot Header */
+/* Audio Icon */
 
-.rcb-chat-header-container {
-	padding: 12px;
-	color: #fff;
-	border-bottom: 1px solid #ccc;
-	display: flex;
-	justify-content: space-between;
-	max-height: 55px;
-}
-
-.rcb-chat-header {
-	display: flex;
-	flex-direction: row;
-}
-
-/* ChatBot Avatar */
-
-.rcb-bot-avatar {
-	background-size: cover;
-	width: 30px;
-	height: 30px;
-	border-radius: 50%;
-	margin-right: 12px;
-}
-
-/* Notification & Audio Icon */
-
-.rcb-notification-icon-on,
-.rcb-notification-icon-off,
 .rcb-audio-icon-on,
 .rcb-audio-icon-off {
 	position: relative;
@@ -40,13 +12,10 @@
 	margin-left: 5px;
 }
 
-.rcb-notification-icon-off,
 .rcb-audio-icon-off {
 	filter: grayscale(100%);
 }
 
-.rcb-notification-icon-on::after,
-.rcb-notification-icon-off::after,
 .rcb-audio-icon-on::after,
 .rcb-audio-icon-off::after {
 	content: "";
@@ -61,9 +30,7 @@
 	opacity: 0;
 	transition: width 0.2s ease-out, height 0.2s ease-out, opacity 0.2s ease-out;
 }
-  
-.rcb-notification-icon-on:hover::after,
-.rcb-notification-icon-off:hover::after,
+
 .rcb-audio-icon-on:hover::after,
 .rcb-audio-icon-off:hover::after {
 	width: 130%;
