+++ src/components/Buttons/EmojiButton/EmojiButton.css
@@ -56,7 +56,7 @@
 
 /* Emoji Picker Popup */
 
-.rcb-emoji-picker-popup {
+.rcb-emoji-button-popup {
 	position: absolute;
 	width: 158px;
 	background-color: #fff;
