+++ src/components/Buttons/NotificationButton/NotificationButton.css
@@ -1,0 +1,39 @@
+/* Notification Icon */
+
+.rcb-notification-icon-on,
+.rcb-notification-icon-off {
+	position: relative;
+	display: inline-block;
+	background-size: cover;
+	width: 30px;
+	height: 30px;
+	border: none;
+	cursor: pointer;
+	margin-left: 5px;
+}
+
+.rcb-notification-icon-off {
+	filter: grayscale(100%);
+}
+
+.rcb-notification-icon-on::after,
+.rcb-notification-icon-off::after {
+	content: "";
+	position: absolute;
+	top: 50%;
+	left: 50%;
+	transform: translate(-50%, -50%);
+	width: 0;
+	height: 0;
+	background-color: rgba(0, 0, 0, 0.1);
+	border-radius: 50%;
+	opacity: 0;
+	transition: width 0.2s ease-out, height 0.2s ease-out, opacity 0.2s ease-out;
+}
+
+.rcb-notification-icon-on:hover::after,
+.rcb-notification-icon-off:hover::after {
+	width: 130%;
+	height: 130%;
+	opacity: 1;
+}
