+++ src/components/Buttons/VoiceButton/VoiceButton.css
