+++ src/constants/Button.ts
@@ -1,0 +1,12 @@
+// identifiers for buttons provided by default in the chatbot
+const Button = {
+	AUDIO_BUTTON: "audio-button",
+	NOTIFICATION_BUTTON: "notification-button",
+	CLOSE_CHAT_BUTTON: "close-chat-button",
+	SEND_MESSAGE_BUTTON: "send-button",
+	VOICE_MESSAGE_BUTTON: "voice-button",
+	FILE_ATTACHMENT_BUTTON: "file-attachment-button",
+	EMOJI_PICKER_BUTTON: "emoji-button"
+}
+
+export { Button };
