+++ src/types/internal/ButtonConfig.ts
@@ -1,0 +1,8 @@
+/**
+ * Defines the button configurations for header, chat input and footer.
+ */
+export type ButtonConfig = {
+	header: JSX.Element[];
+    chatInput: JSX.Element[];
+    footer: JSX.Element[];
+}
